requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==21.4b2
brotlipy==0.7.0
cachetools==5.3.0
chardet==4.0.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
crcmod==1.7
cycler==0.11.0
d2l==0.17.6
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.7
docker-pycreds==0.4.0
easydict==1.13
ecdsa==0.18.0
einops==0.6.0
entrypoints==0.4
fairscale==0.4.6
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.2
fire==0.5.0
flatbuffers==24.3.7
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.4
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.23.0
grpcio==1.53.0
grpcio-tools==1.45.0
h11==0.14.0
h5py==3.8.0
hiq-python==1.1.12
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==2.10
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.9
ipdb==0.13.13
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
lightning-utilities==0.10.0
linkify-it-py==2.0.0
littleutils==0.2.2
llvmlite==0.39.1
lmdb==1.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
mmcv==2.1.0
mmcv-full==1.3.7
mmengine==0.10.3
model-index==0.1.11
mpi4py==3.1.4
multidict==6.0.4
munch==2.5.0
mypy-extensions==1.0.0
networkx==2.6.3
ninja==1.11.1.1
numba==0.56.4
oauthlib==3.2.2
ogb==1.3.6
omegaconf==2.3.0
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.10
oss2==2.17.0
outdated==0.2.2
packaging==23.0
pandas==1.3.5
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
portalocker==2.7.0
prettytable==3.7.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.4
ptflops==0.7
ptyprocess==0.7.0
py-itree==0.0.19
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyblaze==2.3.2
pycave==2.0.5
pycocotools==2.0.7
pycryptodome==3.20.0
pydantic==1.10.7
pydot==1.4.2
pydub==0.25.1
Pygments==2.15.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-mnist==0.7
python-multipart==0.0.6
python-papi==5.5.1.6
pytorch-lightning==1.9.5
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
safetensors==0.4.0
scikit-base==0.4.6
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.7.3
sconf==0.2.5
seaborn==0.12.2
semantic-version==2.10.0
sentencepiece==0.1.98
sentry-sdk==1.19.1
setproctitle==1.3.2
sktime==0.19.2
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
svgwrite==1.4.3
tabulate==0.9.0
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.12
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
tutel @ git+https://github.com/microsoft/tutel@20df39d58745e4a2d4a4dca1350c0684bcdb24b1
typed-ast==1.5.5
typing_extensions==4.7.1
uc-micro-py==1.0.1
uvicorn==0.21.1
virtualenv==20.24.5
wandb==0.14.2
wavedrom==2.0.3.post3
wcwidth==0.2.6
websockets==11.0.1
Werkzeug==2.2.3
wilds==2.0.0
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
yarl==1.8.2
zipp==3.15.0
